Backport code to allow items to be unreachable from ER
This is a backport of the simplified Entrance Randomizer code for
allowing some items the be unreachable.

Theory of operation is rather simple. If the new flag is truned off the
code acts exactly like before, even producing the exact same seeds.

When the flag is turned on, we still obey region and fill rules (so we
dont get accidental keysanity!).  We perform a quick check to see if the
game (with current item placements) is completeable without the item we
are placing. The important data needed for that is already calculated
for checking access. If the item is not required we can skip the access
check, otherwise we do it as normal.