Update options.blade.php
One change, which I can't find the source for, is to change the h4 header to bold. This greatly improves readability across all browsers.

Major changes include:
1) Removal of superfluous information in the difficulty section, to better distinguish between levels.
2) Changing of generic items' (arrows, bombs, rupees) addition/removal in the difficulties section, to be consistent with what is found evident in spoilers.
3) In the Normal Item Pool, Sanctuary Heart Container is combined with the Heart Containers, as is also shown in the
Spoilers, for improved clarity. Generic items have been grouped together at the end of the list, to improve reading,
as was originally done with the Sanctuary Heart Container and Heart Containers.
4) If items contain a form of a hierarchy, they are tiered by noun, then from largest/powerful to smallest/weak. The generic items' pool listing is done this way, as are the shields in the Expert Difficulty.

Minor changes include:
There are minor textual changes and corrections for punctuation, grammar, and consistency.
Removal of an unnecessary </th> tag.
Addition of an alt attribute (W3C-compliant).