Switch h4 to h3 to match the rest of the site
With the update to css bolding h4 tags, this looks out of place.  Other panel-title headers are h3 and look great.  This one looks great without the bold, but the bold ends up applied to it and doesn't look good.  Sorry for not catching this originally when I made the css change.