Move Text Back into Text Table
This moves the text back into the text table. It should be used in
conjunction with the corresponding ROM changes.

The list of strings to blank out were kept separate so as not to discard
the actual translations of the relevant lines. This makes it easy to
remove or restore individual dialog lines on demand.