Rupee Bow sprite fix
Fixed the Sprite corresponding to $7EF340 = 3, or Bow with Silvers (no more arrows), the state that you are put in when you grab a "Bow and Silver Arrows" item, for Rupee Bow mode. This only happens in Customizer, but a hotfix is nice, since it will not affect any other behavior.