Swamp
Adding Access through Hera as a way to skip the Hammer Pegs for the Water Switch.

Not Entry as there is a Key-Door in the Way

Adding Flippers to Entry through Mire

Adding Mire Big Key Placement to consideration for how many Mire SK are needed to reach the Dungeon. (Simmilar to ToH)