add check for Randomizer Enemy Health
There was a Daily Recently which forced progression behind high Health Mini Moldorms in Mini Moldorm Cave before Weapons were available.

This prevents the issue from reappearing by forcing a weapon that can kill stuff before those locations that require you to kill enemies if Enemy Health is randomized. (not caring if the enemies are set to 1 heart or 8)