Remove Flippers Wrap from OWG Logic
Removes Flippers Wrap from OWG by adding the canMirrorWrap flag to them. That flag was chosen due to their effective similarity.

The Glitched Community found this to be a not fitting inclusion for OWG for quite some time now, basically since a Flippers Wrap to Ice Palace was found and included in the Logic.