Add Bunny Pocket Logic to back of Skull Woods
Bunny Pocket is a minor glitch that allows for entry of back of Skull Woods even when Post Aga state prevents Fake Worlds.
This increases the Places where the Moon Pearl can be to the back of Skull Woods and, in the perfect Crystal Layout, Ganons Tower.

Added Unit Tests accordingly.